3.2.3
- Improved Setup Wizard.
- Improved Home.
- Fixed some issues on devices without Google apps installed.
- The text color of dialog buttons will automatically change to white on devices with custom dynamic colors.
- Tiny Vterm will now close itself when using the exit command.
