--- app/src/main/java/com/vectras/vm/SetupQemuActivity.java
@@ -1,985 +1,0 @@
-package com.vectras.vm;
-
-import static android.content.Intent.ACTION_OPEN_DOCUMENT;
-import static android.content.Intent.ACTION_VIEW;
-import static android.view.View.GONE;
-
-import com.termux.app.TermuxActivity;
-import com.termux.app.TermuxService;
-
-import static com.vectras.vm.utils.UIUtils.UIAlert;
-
-import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.app.ProgressDialog;
-import android.content.ClipData;
-import android.content.ClipboardManager;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.net.Uri;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.Environment;
-import android.provider.DocumentsContract;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.Button;
-import android.widget.LinearLayout;
-import android.widget.ProgressBar;
-import android.widget.ScrollView;
-import android.widget.Spinner;
-import android.widget.TextView;
-import android.widget.Toast;
-import android.widget.BaseAdapter;
-
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.preference.PreferenceManager;
-
-import com.google.android.material.button.MaterialButton;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-import com.vectras.qemu.MainSettingsManager;
-import com.vectras.vm.databinding.ActivitySetupQemuBinding;
-import com.vectras.vm.home.HomeActivity;
-import com.vectras.vm.utils.DeviceUtils;
-import com.vectras.vm.utils.DialogUtils;
-import com.vectras.vm.utils.FileUtils;
-import com.vectras.vm.utils.JSONUtils;
-import com.vectras.vm.utils.ListUtils;
-import com.vectras.vm.utils.UIUtils;
-import com.vectras.vm.utils.PermissionUtils;
-import com.vectras.vterm.view.ZoomableTextView;
-
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Objects;
-
-public class SetupQemuActivity extends AppCompatActivity implements View.OnClickListener {
-    ActivitySetupQemuBinding binding;
-    Activity activity;
-    private final String TAG = "SetupQemuActivity";
-    ZoomableTextView vterm;
-    MaterialButton inBtn;
-    ProgressBar progressBar;
-    TextView title;
-
-    LinearLayout linearload;
-    LinearLayout linearcannotconnecttoserver;
-    Button buttontryconnectagain;
-    LinearLayout linearsimplesetupui;
-    LinearLayout linearstartsetup;
-    MaterialButton buttonautosetup;
-    MaterialButton buttonmanualsetup;
-    LinearLayout linearsettingup;
-    TextView textviewsettingup;
-    LinearLayout linearsetupfailed;
-    MaterialButton buttonsetuptryagain;
-    MaterialButton buttonsetupshowlog;
-    TextView textviewshowadvancedsetup;
-    Spinner spinnerselectmirror;
-    LinearLayout linearwelcome;
-    Button buttonletstart;
-
-    private boolean settingup = false;
-    private boolean libprooterror = false;
-
-    private RequestNetwork net;
-    private RequestNetwork.RequestListener _net_request_listener;
-    private String contentJSON = "";
-    private HashMap<String, Object> mmap = new HashMap<>();
-    private String bootstrapfilelink = "";
-    private final ArrayList<HashMap<String, String>> listmapForSelectMirrors = new ArrayList<>();
-    private String selectedMirrorCommand = "echo ";
-    private String selectedMirrorLocation = "";
-    private String downloadBootstrapsCommand = "";
-    private boolean aria2Error = false;
-    private boolean isexecutingCommand = false;
-    private boolean isEdgeServerError = false;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        UIUtils.edgeToEdge(this);
-        binding = ActivitySetupQemuBinding.inflate(getLayoutInflater());
-        setContentView(binding.getRoot());
-        UIUtils.setOnApplyWindowInsetsListener(findViewById(R.id.main));
-        activity = this;
-
-        linearload = findViewById(R.id.linearload);
-        linearcannotconnecttoserver = findViewById(R.id.linearcannotconnecttoserver);
-        buttontryconnectagain = findViewById(R.id.buttontryconnectagain);
-        linearsimplesetupui = findViewById(R.id.linearsimplesetupui);
-        linearstartsetup = findViewById(R.id.linearstartsetup);
-        buttonautosetup = findViewById(R.id.buttonautosetup);
-        buttonmanualsetup = findViewById(R.id.buttonmanualsetup);
-        linearsettingup = findViewById(R.id.linearsettingup);
-        textviewsettingup = findViewById(R.id.textviewsettingup);
-        linearsetupfailed = findViewById(R.id.linearsetupfailed);
-        buttonsetuptryagain = findViewById(R.id.buttonsetuptryagain);
-        buttonsetupshowlog = findViewById(R.id.buttonsetupshowlog);
-        textviewshowadvancedsetup = findViewById(R.id.textviewshowadvancedsetup);
-        spinnerselectmirror = findViewById(R.id.spinnerselectmirror);
-        linearwelcome = findViewById(R.id.linearwelcome);
-        buttonletstart = findViewById(R.id.buttonletstart);
-
-        buttontryconnectagain.setOnClickListener(this);
-        buttonautosetup.setOnClickListener(this);
-        buttonmanualsetup.setOnClickListener(this);
-        buttonsetuptryagain.setOnClickListener(this);
-        buttonsetupshowlog.setOnClickListener(this);
-        textviewshowadvancedsetup.setOnClickListener(this);
-        buttonletstart.setOnClickListener(this);
-
-        progressBar = findViewById(R.id.progressBar);
-
-        vterm = findViewById(R.id.tvTerminalOutput);
-
-        inBtn = findViewById(R.id.btnInstall);
-        title = findViewById(R.id.title);
-        inBtn.setOnClickListener(this);
-        setupSpiner();
-
-        tarPath = getExternalFilesDir("data") + "/data.tar.gz";
-
-        spinnerselectmirror.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
-            @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-                selectedMirrorCommand = Objects.requireNonNull(listmapForSelectMirrors.get(position).get("mirror"));
-                selectedMirrorLocation = Objects.requireNonNull(listmapForSelectMirrors.get(position).get("location"));
-                MainSettingsManager.setSelectedMirror(activity, position);
-            }
-
-            @Override
-            public void onNothingSelected(AdapterView<?> parent) {
-
-            }
-        });
-
-        net = new RequestNetwork(this);
-        _net_request_listener = new RequestNetwork.RequestListener() {
-            @Override
-            public void onResponse(String tag, String response, HashMap<String, Object> responseHeaders) {
-                linearload.setVisibility(GONE);
-                contentJSON = response;
-                if (JSONUtils.isValidFromString(contentJSON)) {
-                    mmap.clear();
-                    mmap = new Gson().fromJson(contentJSON, new TypeToken<HashMap<String, Object>>() {
-                    }.getType());
-                    if (mmap.containsKey("arm64-v8a") && mmap.containsKey("amd64")) {
-                        if (Build.SUPPORTED_ABIS[0].contains("arm64")) {
-                            bootstrapfilelink = Objects.requireNonNull(mmap.get("arm64-v8a")).toString();
-                        } else {
-                            bootstrapfilelink = Objects.requireNonNull(mmap.get("amd64")).toString();
-                        }
-                        downloadBootstrapsCommand = " aria2c -x 4 --async-dns=false --disable-ipv6 --check-certificate=false -o setup.tar.gz " + bootstrapfilelink;
-                        if (!bootstrapfilelink.isEmpty()) {
-                            linearcannotconnecttoserver.setVisibility(GONE);
-                        }
-                    }
-//                    if (AppConfig.needreinstallsystem) {
-//                        AppConfig.needreinstallsystem = false;
-//                        if (!MainSettingsManager.getsetUpWithManualSetupBefore(SetupQemuActivity.this)) {
-//                            if (AppConfig.getSetupFiles().contains("arm64-v8a") || AppConfig.getSetupFiles().contains("x86_64")) {
-//                                setupVectras64();
-//                            } else {
-//                                setupVectras32();
-//                            }
-//                            textviewsettingup.setText(getResources().getString(R.string.reinstalling));
-//                            simpleSetupUIControler(1);
-//                        }
-//                    }
-                }
-            }
-
-            @Override
-            public void onErrorResponse(String tag, String message) {
-                linearload.setVisibility(GONE);
-            }
-        };
-
-        startExtractSystemFiles();
-        setupOnClick();
-    }
-
-    private void startExtractSystemFiles() {
-        String filesDir = activity.getFilesDir().getAbsolutePath();
-        File distroDir = new File(filesDir + "/distro");
-        File binDir = new File(distroDir + "/bin");
-        if (!binDir.exists()) {
-            extractSystemFiles("bootstrap", "");
-            if (Build.SUPPORTED_ABIS[0].contains("64")) {
-                extractSystemFiles("alpine22", "distro");
-            } else {
-                extractSystemFiles("alpine21", "distro");
-            }
-        }
-    }
-
-    public void extractSystemFiles(String fromAsset, String extractTo) {
-        String randomFileName = VMManager.startRamdomVMID();
-        String filesDir = getFilesDir().getAbsolutePath();
-        String abi = getDeviceAbi();
-        //String apkLoaderAssetPath = "bootstrap/loader.apk";
-        String assetPath = fromAsset + "/" + abi + ".tar";
-        String extractedFilePath = filesDir + "/" + randomFileName + ".tar";
-        //String apkLoaderextractedFilePath = TermuxService.PREFIX_PATH + "/libexec/termux-x11/loader.apk";
-        File destDir = new File(filesDir + "/" + extractTo);
-        if (!destDir.exists()) destDir.mkdir();
-
-        ProgressDialog progressDialog = new ProgressDialog(this);
-        progressDialog.setMessage("Extracting data...");
-        progressDialog.setCancelable(false);
-        progressDialog.show();
-
-        new Thread(() -> {
-            boolean isCompleted;
-            String errorMessage = "";
-
-            // Step 1: Copy asset to filesDir
-            if (!copyAssetToFile(assetPath, extractedFilePath)) {
-                errorMessage = "Failed to copy asset file.";
-                isCompleted = false;
-            } else {
-                isCompleted = true;
-            }
-
-            // Step 2: Run tar extraction
-            if (isCompleted) {
-                String[] cmdline = {"tar", "xf", extractedFilePath, "-C", filesDir + "/" + extractTo};
-                Process process = null;
-                try {
-                    process = Runtime.getRuntime().exec(cmdline);
-
-                    // Capture standard error output (stderr)
-                    BufferedReader errorReader =
-                            new BufferedReader(new InputStreamReader(process.getErrorStream()));
-                    StringBuilder errorOutput = new StringBuilder();
-                    String line;
-                    while ((line = errorReader.readLine()) != null) {
-                        errorOutput.append(line).append("\n");
-                    }
-                    errorReader.close();
-
-                    // Wait for the process to complete
-                    int exitCode = process.waitFor();
-
-                    // If there was any output in stderr, treat it as an error
-                    if (exitCode != 0 || errorOutput.length() > 0) {
-                        errorMessage =
-                                errorOutput.toString().isEmpty()
-                                        ? "Unknown error"
-                                        : errorOutput.toString();
-                        isCompleted = false;
-                    }
-                } catch (IOException | InterruptedException e) {
-                    errorMessage = e.getMessage();
-                    isCompleted = false;
-                } finally {
-                    if (process != null) {
-                        process.destroy();
-                    }
-                    //FileUtils.deleteDirectory(apkLoaderextractedFilePath);
-                    //if (!copyAssetToFile(apkLoaderAssetPath, apkLoaderextractedFilePath)) {
-                    //errorMessage = "Failed to copy loader.apk file.";
-                    //}
-                }
-            }
-
-            boolean finalIsCompleted = isCompleted;
-            String finalErrorMessage = errorMessage;
-            runOnUiThread(() -> {
-                progressDialog.dismiss();
-                if (finalIsCompleted) {
-                    Toast.makeText(
-                                    getApplicationContext(),
-                                    R.string.extraction_complete,
-                                    Toast.LENGTH_SHORT)
-                            .show();
-
-                    if (fromAsset.contains("alpine")) {
-                        File rootDir = new File(filesDir + "/distro/root");
-                        if (!rootDir.exists()) rootDir.mkdir();
-
-                        File resolv = new File(filesDir + "/distro/etc/resolv.conf");
-                        resolv.getParentFile().mkdirs();
-                        try (BufferedWriter writer = new BufferedWriter(new FileWriter(resolv))) {
-                            writer.write("nameserver 1.1.1.1\n");
-                            writer.write("nameserver 1.0.0.1\n");
-                        } catch (IOException e) {
-                            Log.e(TAG, "resolv: ", e);
-                        }
-
-                    }
-                } else {
-                    DialogUtils.twoDialog(
-                            this,
-                            getString(R.string.failed),
-                            finalErrorMessage,
-                            getString(R.string.try_again),
-                            getString(R.string.close),
-                            true,
-                            R.drawable.error_96px,
-                            false,
-                            this::startExtractSystemFiles,
-                            this::finish,
-                            null);
-                }
-
-                File extractedTarFile = new File(extractedFilePath);
-                if (extractedTarFile.exists()) extractedTarFile.delete();
-            });
-        }).start();
-    }
-
-    /**
-     * Copies the specified asset to the given file path.
-     */
-    private boolean copyAssetToFile(String assetPath, String outputPath) {
-        try (InputStream in = getAssets().open(assetPath);
-             OutputStream out = new FileOutputStream(outputPath)) {
-            byte[] buffer = new byte[1024];
-            int read;
-            while ((read = in.read(buffer)) != -1) {
-                out.write(buffer, 0, read);
-            }
-            out.flush();
-            return true;
-        } catch (IOException e) {
-            Log.e(TAG, "copyAssetToFile: ", e);
-            return false;
-        }
-    }
-
-    /**
-     * Determines the ABI of the device.
-     */
-    private String getDeviceAbi() {
-        return Build.SUPPORTED_ABIS[0];
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        checkpermissions();
-    }
-
-    public void onClick(View v) {
-        int id = v.getId();
-        if (id == R.id.btnInstall) {
-            File tarGZ = new File(tarPath);
-            if (tarGZ.exists()) {
-                tarGZ.delete();
-            }
-            showAdvancedSetupDialog();
-        } else if (id == R.id.buttonautosetup) {
-            if (AppConfig.getSetupFiles().contains("arm64-v8a") || AppConfig.getSetupFiles().contains("x86_64")) {
-                setupVectras64();
-            } else {
-                setupVectras32();
-            }
-            simpleSetupUIControler(1);
-        } else if (id == R.id.buttonmanualsetup) {
-            Intent intent = new Intent(ACTION_OPEN_DOCUMENT);
-            intent.addCategory(Intent.CATEGORY_OPENABLE);
-            intent.setType("*/*");
-
-            // Optionally, specify a URI for the file that should appear in the
-            // system file picker when it loads.
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Environment.DIRECTORY_DOWNLOADS);
-            }
-
-            startActivityForResult(intent, 1001);
-        } else if (id == R.id.buttonsetuptryagain) {
-            simpleSetupUIControler(0);
-        } else if (id == R.id.buttonsetupshowlog) {
-            linearsimplesetupui.setVisibility(GONE);
-        } else if (id == R.id.textviewshowadvancedsetup) {
-            linearsimplesetupui.setVisibility(GONE);
-            if (linearstartsetup.getVisibility() == View.VISIBLE) {
-                showAdvancedSetupDialog();
-            }
-        } else if (id == R.id.buttontryconnectagain) {
-            linearload.setVisibility(View.VISIBLE);
-            net.startRequestNetwork(RequestNetworkController.GET, AppConfig.bootstrapfileslink, "", _net_request_listener);
-        } else if (id == R.id.buttonletstart) {
-            linearwelcome.setVisibility(GONE);
-            net.startRequestNetwork(RequestNetworkController.GET, AppConfig.bootstrapfileslink, "", _net_request_listener);
-        }
-    }
-
-    String tarPath;
-
-    // Function to append text and automatically scroll to bottom
-    @SuppressLint("SetTextI18n")
-    private void appendTextAndScroll(String textToAdd) {
-        ScrollView scrollView = findViewById(R.id.scrollView);
-
-        // Update the text
-        vterm.append(textToAdd);
-
-        if (textToAdd.contains("xssFjnj58Id")) {
-            isexecutingCommand = false;
-            linearsimplesetupui.setVisibility(GONE);
-            binding.advancedsetup.lnAdvancedsetup.setVisibility(GONE);
-        } else if (textToAdd.contains("libproot.so --help")) {
-            libprooterror = true;
-        } else if (textToAdd.contains("not complete: /root/setup.tar.gz")) {
-            aria2Error = true;
-        } else if (textToAdd.contains("edge") && textToAdd.contains("temporary error")) {
-            isEdgeServerError = true;
-        }
-
-        if (textToAdd.contains("Starting setup...")) {
-            title.setText("Getting ready for you...");
-            textviewsettingup.setText(R.string.getting_ready_for_you_please_don_t_disconnect_the_network);
-        } else if (textToAdd.contains("fetch http")) {
-            title.setText(getString(R.string.connecting_to_mirror_in) + "\n" + selectedMirrorLocation + "...");
-            textviewsettingup.setText(getString(R.string.connecting_to_mirror_in) + "\n" + selectedMirrorLocation + "...");
-        } else if (textToAdd.contains("Installing packages...")) {
-            title.setText(R.string.it_won_t_take_long);
-            textviewsettingup.setText(R.string.completed_10_it_won_t_take_long);
-        } else if (textToAdd.contains("(50/")) {
-            textviewsettingup.setText(R.string.completed_20_it_won_t_take_long);
-        } else if (textToAdd.contains("100/")) {
-            textviewsettingup.setText(R.string.completed_30_it_won_t_take_long);
-        } else if (textToAdd.contains("150/")) {
-            textviewsettingup.setText(R.string.completed_40_it_won_t_take_long);
-        } else if (textToAdd.contains("200/")) {
-            textviewsettingup.setText(R.string.completed_50_it_won_t_take_long);
-        } else if (textToAdd.contains("Downloading Qemu...")) {
-            title.setText(R.string.don_t_disconnect);
-            textviewsettingup.setText(R.string.completed_75_don_t_disconnect);
-        } else if (textToAdd.contains("Installing Qemu...")) {
-            title.setText(R.string.keep_it_up);
-            textviewsettingup.setText(R.string.completed_80_keep_it_up);
-        } else if (textToAdd.contains("qemu-system")) {
-            textviewsettingup.setText(R.string.completed_95_keep_it_up);
-        } else if (textToAdd.contains("Just a sec...")) {
-            title.setText(R.string.almost_there);
-            textviewsettingup.setText(getString(R.string.almost_there));
-        }
-
-        // Scroll to the bottom
-        scrollView.post(() -> scrollView.fullScroll(ScrollView.FOCUS_DOWN));
-    }
-
-    public void onBackPressed() {
-        if (isexecutingCommand) {
-            if (binding.linearsimplesetupui.getVisibility() == GONE)
-                binding.linearsimplesetupui.setVisibility(View.VISIBLE);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    // Method to execute the shell command
-    public void executeShellCommand(String userCommand) {
-        isexecutingCommand = true;
-        new Thread(() -> {
-            try {
-                // Setup the process builder to start PRoot with environmental variables and commands
-                ProcessBuilder processBuilder = new ProcessBuilder();
-
-                // Adjust these environment variables as necessary for your app
-                String filesDir = activity.getFilesDir().getAbsolutePath();
-                String nativeLibDir = activity.getApplicationInfo().nativeLibraryDir;
-
-                File tmpDir = new File(activity.getFilesDir(), "usr/tmp");
-
-                // Setup environment for the PRoot process
-                processBuilder.environment().put("PROOT_TMP_DIR", tmpDir.getAbsolutePath());
-
-                processBuilder.environment().put("HOME", "/root");
-                processBuilder.environment().put("USER", "root");
-                processBuilder.environment().put("PATH", "/bin:/usr/bin:/sbin:/usr/sbin");
-                processBuilder.environment().put("TERM", "xterm-256color");
-                processBuilder.environment().put("TMPDIR", tmpDir.getAbsolutePath());
-                processBuilder.environment().put("SHELL", "/bin/sh");
-
-                String[] prootCommand = {
-                        TermuxService.PREFIX_PATH + "/bin/proot", // PRoot binary path
-                        "--kill-on-exit",
-                        "--link2symlink",
-                        "-0",
-                        "-r", filesDir + "/distro", // Path to the rootfs
-                        "-b", "/dev",
-                        "-b", "/proc",
-                        "-b", "/sys",
-                        "-b", "/sdcard",
-                        "-b", "/storage",
-                        "-b", "/data",
-                        "-w", "/root",
-                        "/bin/sh",
-                        "--login"// The shell to execute inside PRoot
-                };
-
-                processBuilder.command(prootCommand);
-                Process process = processBuilder.start();
-                // Get the input and output streams of the process
-                BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(process.getOutputStream()));
-                BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-                BufferedReader errorReader = new BufferedReader(new InputStreamReader(process.getErrorStream()));
-
-                // Send user command to PRoot
-                writer.write(userCommand);
-                writer.newLine();
-                writer.flush();
-                writer.close();
-
-                // Read the input stream for the output of the command
-                String line;
-                while ((line = reader.readLine()) != null) {
-                    final String outputLine = line;
-                    activity.runOnUiThread(() -> appendTextAndScroll(outputLine + "\n"));
-                }
-
-                // Read any errors from the error stream
-                while ((line = errorReader.readLine()) != null) {
-                    final String errorLine = line;
-                    activity.runOnUiThread(() -> appendTextAndScroll(errorLine + "\n"));
-                }
-
-                // Clean up
-                reader.close();
-                errorReader.close();
-
-                // Wait for the process to finish
-                process.waitFor();
-
-                // Wait for the process to finish
-                int exitValue = process.waitFor();
-
-                // Check if the exit value indicates an error
-                if (exitValue != 0) {
-                    isexecutingCommand = false;
-                    // If exit value is not zero, display a toast message
-                    if (!aria2Error && isEdgeServerError) {
-                        String toastMessage = "Command failed with exit code: " + exitValue;
-                        activity.runOnUiThread(() -> {
-                            appendTextAndScroll("Error: " + toastMessage + "\n");
-                            Toast.makeText(activity, toastMessage, Toast.LENGTH_LONG).show();
-                            uiControllerAdvancedSetup(false);
-                            title.setText(getString(R.string.failed));
-                            simpleSetupUIControler(2);
-                        });
-                    }
-                    if (libprooterror) {
-                        DialogUtils.twoDialog(
-                                this,
-                                getResources().getString(R.string.oops),
-                                getResources().getString(R.string.a_serious_problem_has_occurred),
-                                getString(R.string.join_our_community),
-                                getString(R.string.close),
-                                true, R.drawable.system_update_24px,
-                                true,
-                                () -> {
-                                    Intent intent = new Intent();
-                                    intent.setAction(ACTION_VIEW);
-                                    intent.setData(Uri.parse(AppConfig.community));
-                                    startActivity(intent);
-                                },
-                                null,
-                                null);
-                    } else if (aria2Error && downloadBootstrapsCommand.contains("aria2c")) {
-                        activity.runOnUiThread(() -> {
-                            downloadBootstrapsCommand = " curl -o setup.tar.gz -L " + bootstrapfilelink;
-                            if (AppConfig.getSetupFiles().contains("64")) {
-                                setupVectras64();
-                            } else {
-                                setupVectras32();
-                            }
-                        });
-                    } else if (isEdgeServerError) {
-                        isEdgeServerError = false;
-                        activity.runOnUiThread(() -> {
-                            selectedMirrorCommand += " && sed '/edge/d' /etc/apk/repositories";
-                            if (AppConfig.getSetupFiles().contains("64")) {
-                                setupVectras64();
-                            } else {
-                                setupVectras32();
-                            }
-                        });
-                    }
-                }
-            } catch (IOException | InterruptedException e) {
-                isexecutingCommand = false;
-                // Handle exceptions by printing the stack trace in the terminal output
-                final String errorMessage = e.getMessage();
-                activity.runOnUiThread(() -> {
-                    appendTextAndScroll("Error: " + errorMessage + "\n");
-                    Toast.makeText(activity, "Error executing command: " + errorMessage, Toast.LENGTH_LONG).show();
-                    uiControllerAdvancedSetup(false);
-                    title.setText(getString(R.string.failed));
-                    simpleSetupUIControler(2);
-                });
-            }
-        }).start(); // Execute the command in a separate thread to prevent blocking the UI thread
-    }
-
-    private void setupVectras() {
-        simpleSetupUIControler(1);
-        uiControllerAdvancedSetup(true);
-        progressBar.setVisibility(View.VISIBLE);
-        String neededPkgs = AppConfig.neededPkgs;
-        if (!AppConfig.getSetupFiles().contains("64")) {
-            neededPkgs.replaceAll(" mesa-vulkan-broadcom mesa-vulkan-freedreno mesa-vulkan-panfrost", "");
-        }
-        String cmd = "";
-        cmd += selectedMirrorCommand + ";";
-        String cmd2 = "set -e;" +
-                " echo \"Starting setup...\";" +
-                " apk update;" +
-                " echo \"Installing packages...\";" +
-                " apk add " + neededPkgs + ";" +
-                " echo \"Installing Qemu...\";" +
-                " tar -xzvf " + tarPath + " -C /;" +
-                " rm " + tarPath + ";" +
-                " echo \"Just a sec...\";" +
-                " apk add qemu-audio-sdl pulseaudio;" +
-                " echo export PULSE_SERVER=127.0.0.1 >> /etc/profile;" +
-                " mkdir -p ~/.vnc && echo -e \"555555\\n555555\" | vncpasswd -f > ~/.vnc/passwd && chmod 0600 ~/.vnc/passwd;" +
-                " chmod 775 /usr/local/bin/*;" +
-                " echo \"installation successful! xssFjnj58Id\"";
-
-        executeShellCommand(cmd);
-        executeShellCommand(cmd2);
-
-        binding.advancedsetup.tvCommandsetup.setText(cmd + cmd2);
-    }
-
-    private void setupVectras32() {
-        aria2Error = false;
-        uiControllerAdvancedSetup(true);
-        progressBar.setVisibility(View.VISIBLE);
-        String cmd = "";
-        cmd += selectedMirrorCommand + ";";
-        String cmd2 = "set -e;" +
-                " echo \"Starting setup...\";" +
-                " apk update;" +
-                " echo \"Installing packages...\";" +
-                " apk add " + AppConfig.neededPkgs32bit + ";" +
-                //" tar -xzvf " + tarPath + " -C /;" +
-                " echo \"Installing Qemu...\";" +
-                " apk add qemu-system-x86_64 qemu-system-ppc qemu-system-i386 qemu-system-aarch64 qemu-pr-helper qemu-img qemu-audio-sdl pulseaudio mesa-dri-gallium;" +
-                " echo \"Just a sec...\";" +
-                " echo export PULSE_SERVER=127.0.0.1 >> /etc/profile;" +
-                //" rm " + tarPath + ";" +
-                " mkdir -p ~/.vnc && echo -e \"555555\\n555555\" | vncpasswd -f > ~/.vnc/passwd && chmod 0600 ~/.vnc/passwd;" +
-                " echo \"installation successful! xssFjnj58Id\"";
-
-        executeShellCommand(cmd);
-        executeShellCommand(cmd2);
-
-        binding.advancedsetup.tvCommandsetup.setText(cmd + cmd2);
-    }
-
-    private void setupVectras64() {
-        aria2Error = false;
-        uiControllerAdvancedSetup(true);
-        progressBar.setVisibility(View.VISIBLE);
-        String cmd = "";
-        cmd += selectedMirrorCommand + ";";
-        String cmd2 = "set -e;" +
-                " echo \"Starting setup...\";" +
-                " apk update;" +
-                " echo \"Installing packages...\";" +
-                " apk add " + AppConfig.neededPkgs + ";" +
-                " echo \"Downloading Qemu...\";" +
-                downloadBootstrapsCommand + ";" +
-                " echo \"Installing Qemu...\";" +
-                " tar -xzvf setup.tar.gz -C /;" +
-                " rm setup.tar.gz;" +
-                " echo \"Just a sec...\";" +
-                " apk add qemu-audio-sdl pulseaudio;" +
-                " echo export PULSE_SERVER=127.0.0.1 >> /etc/profile;" +
-                " mkdir -p ~/.vnc && echo -e \"555555\\n555555\" | vncpasswd -f > ~/.vnc/passwd && chmod 0600 ~/.vnc/passwd;" +
-                " chmod 775 /usr/local/bin/*;" +
-                " echo \"installation successful! xssFjnj58Id\"";
-
-        executeShellCommand(cmd);
-        executeShellCommand(cmd2);
-
-        binding.advancedsetup.tvCommandsetup.setText(cmd + cmd2);
-    }
-
-    private void checkabi() {
-        if (!AppConfig.getSetupFiles().contains("arm64-v8a")) {
-            if (!AppConfig.getSetupFiles().contains("x86_64")) {
-                UIUtils.oneDialog(getResources().getString(R.string.warning), getResources().getString(R.string.cpu_not_support_64), true, false, activity);
-            }
-        }
-    }
-
-    private void checkpermissions() {
-        if (PermissionUtils.storagepermission(activity, true)) {
-            if (!settingup) {
-                settingup = true;
-                checkabi();
-
-                File tarGZ = new File(tarPath);
-                if (tarGZ.exists()) {
-                    setupVectras();
-                } else {
-                    if (linearsimplesetupui.getVisibility() == GONE) {
-                        showAdvancedSetupDialog();
-                    }
-                }
-            }
-
-            if (DeviceUtils.isStorageLow(this, false)) {
-                DialogUtils.oneDialog(this,
-                        getResources().getString(R.string.oops),
-                        getResources().getString(R.string.not_enough_storage_to_set_up_content),
-                        getResources().getString(R.string.ok),
-                        true,
-                        R.drawable.warning_48px,
-                        true,
-                        null,
-                        () -> {
-                            if (DeviceUtils.isStorageLow(this, false)) finish();
-                        });
-            }
-        }
-    }
-
-    private void simpleSetupUIControler(int status) {
-        if (status == 0) {
-            linearstartsetup.setVisibility(View.VISIBLE);
-            linearsettingup.setVisibility(GONE);
-            linearsetupfailed.setVisibility(GONE);
-        } else if (status == 1) {
-            linearstartsetup.setVisibility(GONE);
-            linearsettingup.setVisibility(View.VISIBLE);
-            linearsetupfailed.setVisibility(GONE);
-        } else if (status == 2) {
-            linearstartsetup.setVisibility(GONE);
-            linearsettingup.setVisibility(GONE);
-            linearsetupfailed.setVisibility(View.VISIBLE);
-        }
-    }
-
-    public String getPath(Uri uri) {
-        return FileUtils.getPath(this, uri);
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent ReturnedIntent) {
-        super.onActivityResult(requestCode, resultCode, ReturnedIntent);
-        if (requestCode == 1001 && resultCode == RESULT_OK) {
-            Uri content_describer = ReturnedIntent.getData();
-            File selectedFilePath = new File(getPath(content_describer));
-            ProgressBar loading = progressBar;
-            String abi = Build.SUPPORTED_ABIS[0];
-            if (selectedFilePath.toString().endsWith(abi + ".tar.gz")) {
-                simpleSetupUIControler(1);
-                loading.setVisibility(View.VISIBLE);
-                new Thread(() -> {
-                    FileInputStream File;
-                    try {
-                        assert content_describer != null;
-                        File = (FileInputStream) getContentResolver().openInputStream(content_describer);
-                    } catch (FileNotFoundException e) {
-                        throw new RuntimeException(e);
-                    }
-                    try {
-                        try {
-                            try (OutputStream out = new FileOutputStream(tarPath)) {
-                                // Transfer bytes from in to out
-                                byte[] buf = new byte[1024];
-                                int len;
-                                while (true) {
-                                    assert File != null;
-                                    if (!((len = File.read(buf)) > 0)) break;
-                                    out.write(buf, 0, len);
-                                }
-                            }
-                        } finally {
-                            Runnable runnable = () -> {
-                                loading.setVisibility(GONE);
-                                setupVectras();
-                                MainSettingsManager.setsetUpWithManualSetupBefore(SetupQemuActivity.this, true);
-                            };
-                            activity.runOnUiThread(runnable);
-                            assert File != null;
-                            File.close();
-                        }
-                    } catch (IOException e) {
-                        Runnable runnable = () -> {
-                            loading.setVisibility(GONE);
-                            UIAlert(activity, e.toString(), "error");
-                        };
-                        activity.runOnUiThread(runnable);
-                    }
-                }).start();
-            } else {
-                if (linearsimplesetupui.getVisibility() == GONE) {
-                    showAdvancedSetupDialog();
-                }
-                UIAlert(activity, "INVALID FILE", "please select vectras-vm-" + abi + ".tar.gz file");
-            }
-        } else if (linearsimplesetupui.getVisibility() == GONE) {
-            showAdvancedSetupDialog();
-        }
-    }
-
-    private void setupSpiner() {
-        ListUtils.setupMirrorListForListmap(listmapForSelectMirrors);
-
-        spinnerselectmirror.setAdapter(new SpinnerSelectMirrorAdapter(listmapForSelectMirrors));
-        spinnerselectmirror.setSelection(MainSettingsManager.getSelectedMirror(activity));
-    }
-
-    private void setupOnClick() {
-        LinearLayout linearcommunity = findViewById(R.id.linearcommunity);
-        LinearLayout lineardonate = findViewById(R.id.lineardonate);
-
-        Button dontjointtelegrambutton = findViewById(R.id.dontjointtelegrambutton);
-        dontjointtelegrambutton.setOnClickListener(v -> linearcommunity.setVisibility(GONE));
-
-        Button jointelegrambutton = findViewById(R.id.jointelegrambutton);
-        jointelegrambutton.setOnClickListener(v -> {
-            linearcommunity.setVisibility(GONE);
-            Intent intent = new Intent(ACTION_VIEW, Uri.parse(AppConfig.telegramLink));
-            startActivity(intent);
-            //Don't show join Telegram dialog again
-            SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(activity);
-            SharedPreferences.Editor edit = prefs.edit();
-            edit.putBoolean("tgDialog", true);
-            edit.apply();
-        });
-
-        Button dontdonatebutton = findViewById(R.id.dontdonatebutton);
-        dontdonatebutton.setOnClickListener(v -> lineardonate.setVisibility(GONE));
-
-        Button donatebuuton = findViewById(R.id.donatebutton);
-        donatebuuton.setOnClickListener(v -> {
-            lineardonate.setVisibility(GONE);
-            Intent intent = new Intent(ACTION_VIEW, Uri.parse(AppConfig.patreonLink));
-            startActivity(intent);
-        });
-
-        Button done100 = findViewById(R.id.done100);
-        done100.setOnClickListener(v -> {
-            if (!AppConfig.needreinstallsystem) {
-                startActivity(new Intent(SetupQemuActivity.this, HomeActivity.class));
-            } else {
-                AppConfig.needreinstallsystem = false;
-            }
-            finish();
-        });
-
-        binding.advancedsetup.ivClose.setOnClickListener(v -> linearsimplesetupui.setVisibility(View.VISIBLE));
-        binding.advancedsetup.ivOpenterminal.setOnClickListener(v -> {
-            if (DeviceUtils.is64bit()) {
-                startActivity(new Intent(this, TermuxActivity.class));
-            } else {
-                com.vectras.vterm.TerminalBottomSheetDialog VTERM = new com.vectras.vterm.TerminalBottomSheetDialog(this);
-                VTERM.showVterm();
-            }
-        });
-
-        binding.advancedsetup.ivCopycommandsetup.setOnClickListener(v -> copyToClipboard(binding.advancedsetup.tvCommandsetup.getText().toString()));
-    }
-
-    private void uiControllerAdvancedSetup(boolean isStartSetup) {
-        binding.advancedsetup.lnBntinstall.setVisibility(isStartSetup ? GONE : View.VISIBLE);
-        binding.advancedsetup.progressBar.setVisibility(isStartSetup ? View.VISIBLE : GONE);
-    }
-
-    private void showAdvancedSetupDialog() {
-        DialogUtils.twoDialog(
-                this,
-                getResources().getString(R.string.bootstrap_required),
-                getResources().getString(R.string.you_can_choose_between_auto_download_and_setup_or_manual_setup_by_choosing_bootstrap_file),
-                getString(R.string.auto_setup),
-                getString(R.string.manual_setup),
-                true, R.drawable.system_update_24px,
-                false,
-                () -> {
-                    //startDownload();
-                    if (AppConfig.getSetupFiles().contains("arm64-v8a") || AppConfig.getSetupFiles().contains("x86_64")) {
-                        setupVectras64();
-                    } else {
-                        setupVectras32();
-                    }
-                    simpleSetupUIControler(1);
-                },
-                () -> {
-                    Intent intent = new Intent(ACTION_OPEN_DOCUMENT);
-                    intent.addCategory(Intent.CATEGORY_OPENABLE);
-                    intent.setType("*/*");
-
-                    // Optionally, specify a URI for the file that should appear in the
-                    // system file picker when it loads.
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                        intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Environment.DIRECTORY_DOWNLOADS);
-                    }
-
-                    startActivityForResult(intent, 1001);
-                },
-                null);
-    }
-
-    public class SpinnerSelectMirrorAdapter extends BaseAdapter {
-
-        ArrayList<HashMap<String, String>> _data;
-
-        public SpinnerSelectMirrorAdapter(ArrayList<HashMap<String, String>> _arr) {
-            _data = _arr;
-        }
-
-        @Override
-        public int getCount() {
-            return _data.size();
-        }
-
-        @Override
-        public HashMap<String, String> getItem(int _index) {
-            return _data.get(_index);
-        }
-
-        @Override
-        public long getItemId(int _index) {
-            return _index;
-        }
-
-        @Override
-        public View getView(final int _position, View _v, ViewGroup _container) {
-            LayoutInflater _inflater = getLayoutInflater();
-            View _view = _v;
-            if (_view == null) {
-                _view = _inflater.inflate(R.layout.simple_layout_for_spiner, null);
-            }
-
-            final TextView textViewLocation = _view.findViewById(R.id.textViewLocation);
-
-            textViewLocation.setText(Objects.requireNonNull(_data.get(_position).get("location")));
-
-            return _view;
-        }
-    }
-
-    private void copyToClipboard(String text) {
-        ClipboardManager clipboard = (ClipboardManager) getSystemService(CLIPBOARD_SERVICE);
-        ClipData clip = ClipData.newPlainText("Vectras VM", text);
-        clipboard.setPrimaryClip(clip);
-        Toast.makeText(this, getString(R.string.copied), Toast.LENGTH_SHORT).show();
-    }
-}
