+++ app/src/main/java/com/vectras/vm/home/HomeActivity.java
@@ -135,7 +135,7 @@ public class HomeActivity extends AppCompatActivity implements RomStoreFragment.
 //        UIUtils.setOnApplyWindowInsetsListenerTop(bindingContent.main);
 //        UIUtils.setOnApplyWindowInsetsListenerLeftOnly(binding.navView);
         UIUtils.setOnApplyWindowInsetsListenerBottomOnly(binding.rvRomstoresearch);
-//        UIUtils.setOnApplyWindowInsetsListenerBottomOnly(binding.lnSearchempty);
+        UIUtils.setOnApplyWindowInsetsListenerBottomOnly(binding.lnSearchempty);
 
         initialize(bundle);
     }
