+++ app/src/main/res/drawable/upcoming_96px.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="96dp"
+    android:height="96dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960"
+    android:tint="?attr/colorControlNormal">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M160,840Q127,840 103.5,816.5Q80,793 80,760L80,560Q80,527 103.5,503.5Q127,480 160,480L360,480Q360,530 395,565Q430,600 480,600Q530,600 565,565Q600,530 600,480L800,480Q833,480 856.5,503.5Q880,527 880,560L880,760Q880,793 856.5,816.5Q833,840 800,840L160,840ZM160,760L800,760Q800,760 800,760Q800,760 800,760L800,560Q800,560 800,560Q800,560 800,560L664,560Q639,615 589.5,647.5Q540,680 480,680Q420,680 370.5,647.5Q321,615 296,560L160,560Q160,560 160,560Q160,560 160,560L160,760Q160,760 160,760Q160,760 160,760ZM704,432L648,376L790,234L846,290L704,432ZM256,432L114,290L170,234L312,376L256,432ZM440,320L440,120L520,120L520,320L440,320ZM160,760Q160,760 160,760Q160,760 160,760L160,760Q160,760 160,760Q160,760 160,760L296,760Q321,760 370.5,760Q420,760 480,760Q540,760 589.5,760Q639,760 664,760L800,760Q800,760 800,760Q800,760 800,760L800,760Q800,760 800,760Q800,760 800,760L160,760Z"/>
+</vector>
